test: flatten fs test module and isolate allowlist regressions for Windows/CodeScene (#1224)
